Moved the implemented functions to a new file in utils, and added an example to the quantiled thresholding function.
