+++ petpal/preproc/image_operations_4d.py
@@ -26,12 +26,10 @@ import os
 import pathlib
 import datetime
 import tempfile
-import re
 import ants
 import nibabel
 import numpy as np
 from scipy.ndimage import center_of_mass
-from sklearn.decomposition import PCA
 
 from ..utils.useful_functions import weighted_series_sum, check_physical_space_for_ants_image_pair
 from ..utils import image_io, math_lib
@@ -604,316 +602,6 @@ def extract_roi_voxel_tacs_from_image_using_mask(input_image: ants.core.ANTsImag
     if verbose:
         print(f"(ImageOps): Output TACs have shape {out_voxels.shape}")
     return out_voxels
-
-
-def calculate_temporal_pca_from_image_using_mask(input_image: ants.core.ANTsImage,
-                                                 mask_image: ants.core.ANTsImage,
-                                                 num_components: int = 3,
-                                                 svd_solver: str = 'full',
-                                                 whiten: bool = True,
-                                                 **sklearn_pca_kwargs) -> tuple[PCA, np.ndarray]:
-    """
-    Perform temporal PCA (Principal Component Analysis) on a 4D PET image using a specified mask.
-
-    This function analyzes the temporal dynamics of voxel time-activity curves (TACs) extracted
-    from the region specified by the input mask. PCA reduces the dimensionality of the dataset,
-    extracting a set of temporal features ranked by their contribution to variability across
-    voxels within the specified region. This approach effectively identifies the most prominent
-    temporal patterns in the PET image's dynamic data.
-
-    Args:
-        input_image (ants.core.ANTsImage):
-            The 4D PET image for temporal PCA analysis. The image must have 3 spatial dimensions
-            and one temporal dimension.
-        mask_image (ants.core.ANTsImage):
-            A binary mask representing the region of interest (ROI) in the input image.
-            Only voxels within this mask are included in the PCA computation. The mask and input
-            image must have the same physical dimensions.
-        num_components (int, optional):
-            The number of principal components to compute. Defaults to 3.
-        svd_solver (str, optional): The type of SVD solver to use for PCA. Defaults to `"full"`.
-        whiten (bool, optional):
-            If True, the output features are normalized to have unit variance. This is useful
-            for further statistical analysis of the temporal components. Defaults to True.
-        **sklearn_pca_kwargs:
-            Additional keyword arguments to customize the PCA behavior using scikit-learn's
-            `PCA` class (e.g., `random_state`, `tol`, etc.).
-
-    Returns:
-        tuple[PCA, np.ndarray]: Fitted PCA model object and voxel TACs projected onto the principal components.
-
-    Raises:
-        AssertionError: The input image is not 4D.
-        AssertionError: The input image and mask image are not in the same physical space or dimensions.
-
-    Notes:
-        - Ensure that the input image and mask image are pre-registered and aligned, with identical
-          spatial dimensions.
-        - The PCA algorithm identifies patterns across the temporal dynamics while minimizing loss
-          of information.
-        - Whitening normalizes the data, which can be important if further statistical operations
-          (e.g., clustering) are planned.
-
-    See Also:
-        - :func:`extract_roi_voxel_tacs_from_image_using_mask`: Extracts region-specific voxel-level
-          time-activity curves (TACs) needed for this PCA analysis.
-        - :class:`sklearn.decomposition.PCA`: Underlying PCA implementation used in this function.
-    """
-    assert len(input_image.shape) == 4, "Input image must be 4D."
-    assert check_physical_space_for_ants_image_pair(input_image, mask_image), (
-        "Images must have the same physical dimensions.")
-    mask_voxels = extract_roi_voxel_tacs_from_image_using_mask(input_image=input_image,
-                                                               mask_image=mask_image)
-    pca_obj = PCA(n_components=num_components, svd_solver=svd_solver, whiten=whiten, **sklearn_pca_kwargs)
-    pca_obj.fit(mask_voxels)
-    return pca_obj, np.ascontiguousarray(pca_obj.fit_transform(mask_voxels), float)
-
-
-def extract_temporal_pca_components_from_image_using_mask(input_image: ants.core.ANTsImage,
-                                                          mask_image: ants.core.ANTsImage,
-                                                          num_components: int = 3,
-                                                          svd_solver: str = 'full',
-                                                          whiten: bool = True,
-                                                          **sklearn_pca_kwargs) -> np.ndarray:
-    """
-    Extract principal components from temporal PCA on a 4D PET image using a specified mask.
-
-    This function performs temporal Principal Component Analysis (PCA) on a 4D input image,
-    focusing on the region defined by the mask. It returns only the PCA components
-    (i.e., the eigenvectors that describe temporal variation) from the fitted PCA model.
-    These components might help you identify the primary kinetic components over the mask.
-
-    Args:
-        input_image (ants.core.ANTsImage):
-            The 4D PET image for temporal PCA analysis. The image should have 3 spatial
-            dimensions and one temporal dimension.
-        mask_image (ants.core.ANTsImage):
-            A binary mask defining the region of interest (ROI) in the input image.
-            Only voxels within this mask will be considered during PCA computation.
-            The mask and the input image must share the same physical and spatial dimensions.
-        num_components (int, optional):
-            The number of principal components to compute. Defaults to 3.
-        svd_solver (str, optional):
-            The type of SVD solver to use for PCA. Defaults to `"full"`.
-        whiten (bool, optional):
-            If set to `True`, the resulting components are scaled to have unit variance.
-            Defaults to `True`.
-        **sklearn_pca_kwargs:
-            Additional keyword arguments for customizing PCA behavior via scikit-learn's
-            `PCA` class, such as `random_state`, `tol`, etc.
-
-    Returns:
-        np.ndarray:
-            A 2D array of shape `(num_components, num_timepoints)` representing the PCA
-            components of temporal features. Each row corresponds to a principal component
-            in the temporal domain.
-
-    Raises:
-        AssertionError:
-            Raised if the input image is not 4D, or if the mask is not in the same
-            physical space as the input image.
-        ValueError:
-            Raised if the number of components exceeds the valid limits given the input
-            data size.
-
-    Notes:
-        - This function is a higher-level utility that focuses solely on extracting
-          PCA components for temporal dynamics, assuming the underlying computation is
-          performed via a separate PCA implementation.
-        - Ensure that the input and mask images are in the same coordinate space
-          (aligned and registered) to avoid errors or mismatched results.
-
-    See Also:
-        - :func:`calculate_temporal_pca_from_image_using_mask`: Performs full temporal PCA
-          and returns both the fitted PCA model and the projections of voxel time-activity
-          curves.
-        - :class:`sklearn.decomposition.PCA`: Core class for performing Principal
-          Component Analysis in scikit-learn.
-
-    """
-    pca_obj, _ = calculate_temporal_pca_from_image_using_mask(input_image=input_image,
-                                                           mask_image=mask_image,
-                                                           num_components=num_components,
-                                                           svd_solver=svd_solver,
-                                                           whiten=whiten,
-                                                           **sklearn_pca_kwargs)
-    return pca_obj.components_
-
-
-def extract_temporal_pca_projection_from_image_using_mask(input_image: ants.core.ANTsImage,
-                                                          mask_image: ants.core.ANTsImage,
-                                                          num_components: int = 3,
-                                                          svd_solver: str = 'full',
-                                                          whiten: bool = True,
-                                                          **sklearn_pca_kwargs) -> np.ndarray:
-    """
-    Compute and extract the temporal PCA projections for a masked region in a 4D PET image.
-
-    This function carries out temporal Principal Component Analysis (PCA) on a 4D input image,
-    focusing on the region defined by the mask. It returns the projections (i.e., the representation
-    of each voxel in the reduced-dimensional space of principal components).
-
-    Args:
-        input_image (ants.core.ANTsImage):
-            A 4D PET image for temporal PCA analysis. The image should have 3 spatial
-            dimensions and one temporal dimension.
-        mask_image (ants.core.ANTsImage):
-            A binary mask specifying the region of interest (ROI) to include in the PCA analysis.
-            The mask must have the same physical and spatial dimensions as the input image.
-        num_components (int, optional):
-            The number of principal components to compute. Determines the dimensionality of
-            the reduced output space. Defaults to 3.
-        svd_solver (str, optional):
-            Determines the solver used for Singular Value Decomposition (SVD) during PCA.
-            Defaults to `"full"`.
-        whiten (bool, optional):
-            If set to `True`, components are scaled to have unit variance. This is often
-            useful when the output is used as input to further statistical or machine learning models.
-            Defaults to `True`.
-        **sklearn_pca_kwargs:
-            Additional keyword arguments for scikit-learn's `PCA` class, such as specification
-            of `random_state`, `tol`, etc.
-
-    Returns:
-        np.ndarray:
-            A 2D array of shape `(num_voxels, num_components)` containing the temporal PCA
-            projections for each voxel within the mask. Each row represents the reduced-dimensional
-            representation of a voxel's TAC within the selected components.
-
-    Raises:
-        AssertionError:
-            Raised if the `input_image` is not 4D, or if the physical dimensions of the
-            `mask_image` and `input_image` are not identical.
-        ValueError:
-            Raised if the number of PCA components exceeds the number of voxels included
-            in the mask.
-
-    Notes:
-        - The projections represent the linear combination of the temporal principal components
-          for each voxel, condensed into a reduced space.
-        - Ensure the input and mask are registered to the same coordinate space and dimensions
-          to avoid misalignment issues.
-        - For detailed analysis, consider reviewing the explained variance ratios from the
-          corresponding PCA computation.
-
-    See Also:
-        - :func:`calculate_temporal_pca_from_image_using_mask`: Performs the underlying PCA
-          calculation and provides both PCA projections and the PCA model.
-        - :class:`sklearn.decomposition.PCA`: Core implementation of the PCA used in this function.
-    """
-    _, transformed_voxels = calculate_temporal_pca_from_image_using_mask(input_image=input_image,
-                                                           mask_image=mask_image,
-                                                           num_components=num_components,
-                                                           svd_solver=svd_solver,
-                                                           whiten=whiten,
-                                                           **sklearn_pca_kwargs)
-    return transformed_voxels
-
-
-def extract_temporal_pca_quantile_thresholded_tac_vals_from_image_using_mask(input_image: ants.core.ANTsImage,
-                                                                             mask_image: ants.core.ANTsImage,
-                                                                             num_components: int = 3,
-                                                                             threshold_components: list[int] | None = None,
-                                                                             quantiles: np.ndarray = np.asarray(
-                                                                                     [0.5, 0.75, 0.9,
-                                                                                      0.975]),
-                                                                             direction: str = '>',
-                                                                             **sklearn_pca_kwargs) -> np.ndarray:
-    """
-    Extract quantile-thresholded time-activity curve (TAC) values for temporal PCA components from a 4D PET image.
-
-    This function computes time-activity curve (TAC) statistics for masked regions in a 4D PET image, based on
-    Principal Component Analysis (PCA). It thresholds the PCA projections (for provided selected components) using
-    specified quantiles, and calculates the mean and standard deviation of the TACs for voxels exceeding the
-    thresholds. The output includes both the TAC mean and standard deviation across all components and quantile
-    thresholds.
-
-    Args:
-        input_image (ants.core.ANTsImage):
-            A 4D PET image.
-        mask_image (ants.core.ANTsImage):
-            A binary mask specifying the region of interest (ROI) in the image. Only voxels within this mask
-            are considered for the PCA analysis and TAC computation. This image is assumed to be aligned
-            to the input image.
-        num_components (int, optional):
-            The number of PCA components to use for reducing the dataset's dimensionality. Defaults to 3.
-        threshold_components (list[int] | None, optional):
-            A list of component indices (from 0 to `num_components - 1`) to use for thresholding.
-            Thresholding is applied to these PCA components using the given quantiles. If not set, defaults to `[0]`.
-        quantiles (np.ndarray, optional):
-            An array of quantiles (values between 0 and 1) to compute thresholds for the PCA projections.
-            Only voxels with PCA projection values above the corresponding quantile thresholds are included
-            when calculating TAC statistics. Defaults to `[0.5, 0.75, 0.9, 0.975]`.
-        direction (str, optional):
-            An optional argument to switch from ``>`` to `<`. Defaults to ``>``.
-        **sklearn_pca_kwargs:
-            Additional keyword arguments passed to customize scikit-learn's PCA class (e.g., `svd_solver`,
-            `random_state`, etc.).
-
-    Returns:
-        np.ndarray:
-            A 2-element array containing `out_vals` (*np.ndarray*),  a 3D array of shape
-            `(num_components, num_quantiles, num_timepoints)`, containing the mean TAC values of
-            voxels exceeding the quantile thresholds for each component, and `out_stds` (*np.ndarray*),
-            a 3D array of the same shape as `out_vals`, containing the standard deviation of the TAC
-            values under the same conditions.
-
-    Raises:
-        AssertionError: `threshold_components` contains invalid indices (e.g., negative or greater than `num_components - 1`).
-        AssertionError: Quantiles are not within the [0, 1] range.
-
-    Notes:
-        - This function requires a properly aligned input image and mask image (both should share the same
-          dimensions in physical and voxel space).
-        - Thresholding PCA projections using quantiles isolates specific temporal dynamics for analysis,
-          which can provide insight into dominant trends in the time-activity data.
-        - Voxels not exceeding the quantile thresholds are excluded from the mean and standard deviation
-          calculations.
-
-    See Also:
-        - :func:`extract_temporal_pca_projection_from_image_using_mask`: Computes PCA projections used as input
-          for thresholding in this function.
-        - :func:`extract_roi_voxel_tacs_from_image_using_mask`: Extracts voxel-level TACs for regions of interest
-          in the input image, used in TAC calculations here.
-        - :class:`sklearn.decomposition.PCA`: Core PCA implementation used in this analysis.
-    """
-
-    if threshold_components is None:
-        threshold_components = [0]
-
-    assert min(threshold_components) >= 0, "Threshold components must be integers >= 0."
-    assert max(threshold_components) < num_components, "Threshold components must be < num_components."
-    assert np.min(quantiles) >= 0, "Quantiles must be >= 0."
-    assert np.max(quantiles) <= 1, "Quantiles must be <= 1."
-
-    voxels_pca_projs = extract_temporal_pca_projection_from_image_using_mask(input_image=input_image,
-                                                                            mask_image=mask_image,
-                                                                            num_components=num_components,
-                                                                            **sklearn_pca_kwargs)
-    voxels_selected_pca = voxels_pca_projs[:, threshold_components]
-
-    thresholds = np.quantile(voxels_selected_pca, quantiles, axis=0)
-    thresholds = np.expand_dims(thresholds.T, axis=0)
-
-    voxels_selected_pca = voxels_selected_pca[..., None]
-    valid_voxels_mask = (voxels_selected_pca < thresholds) if direction == '<' else (voxels_selected_pca > thresholds)
-    valid_voxels_mask = np.expand_dims(valid_voxels_mask, axis=2)
-
-    roi_voxels = extract_roi_voxel_tacs_from_image_using_mask(input_image=input_image, mask_image=mask_image)
-    num_voxels, num_frames = roi_voxels.shape
-    num_comps, num_quantiles = len(threshold_components), len(quantiles)
-
-    roi_voxels = np.expand_dims(roi_voxels, axis=(1,-1))
-    roi_voxels = np.broadcast_to(roi_voxels, (num_voxels, num_comps, num_frames, num_quantiles))
-
-    masked_voxels = np.where(valid_voxels_mask, roi_voxels, np.nan)
-    tacs_mean = np.nanmean(masked_voxels, axis=0)
-    tacs_std = np.nanstd(masked_voxels, axis=0)
-
-    tacs_mean = np.moveaxis(tacs_mean, 1, -1)
-    tacs_std = np.moveaxis(tacs_std, 1, -1)
-
-    return np.asarray([tacs_mean, tacs_std])
 
 
 class SimpleAutoImageCropper(object):
